# src/main/resources/openapi/bazar-space-openapi.yaml
openapi: 3.1.0
info:
  title: Bazar space openapi
  version: v1
paths:
  /space:
    get:
      tags:
        - space-controller
      operationId: getAllSpaces
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/GetSpacesResponse"
    post:
      tags:
        - space-admin-controller
      operationId: createSpace
      parameters:
        - in: query
          name: name
          description: space name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: '#/components/schemas/GetSpaceDto'
  /space/user:
    post:
      tags:
        - space-admin-controller
      operationId: addUserToSpace
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddUserToSpaceDtoRequest"
        required: true
      responses:
        "200":
          description: OK
  /spaces/{spaceId}/users:
    get:
      tags:
        - space-controller
      operationId: getAllUsersInSpaceEnriched
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/GetUsersInSpaceResponse"
  /space/{spaceId}:
    delete:
      tags:
        - space-admin-controller
      operationId: deleteSpace
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
    patch:
      tags:
        - space-admin-controller
      operationId: patchSpace
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "application/json":
              schema:
                $ref: "#/components/schemas/GetSpaceDto"
  /space/{spaceId}/user/{userId}:
    delete:
      tags:
        - space-admin-controller
      operationId: deleteUserFromSpace
      parameters:
        - name: spaceId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
components:
  schemas:
    AddUserToSpaceDtoRequest:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        spaceId:
          type: integer
          format: int64
      required:
        - userId
        - spaceId
    GetSpaceDto:
      description: space info for get requests
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    GetSpacesResponse:
      type: object
      properties:
        spaces:
          type: array
          items:
            $ref: '#/components/schemas/GetSpaceDto'

    GetUsersInSpaceResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserInSpaceDto'

    UserInSpaceDto:
      type: object
      required:
        - userId
        - spaceId
        - userName
        - firstName
        - lastName
      properties:
        userId:
          type: string
          format: uuid
          description: User ID
        spaceId:
          type: integer
          format: int64
          description: Space ID
        userName:
          type: string
          description: userName from bazar-persona/users
        firstName:
          type: string
          description: firstName from bazar-persona/users
        lastName:
          type: string
          description: lastName from bazar-persona/users
        role:
          type: object
          $ref: '#/components/schemas/SimpleRoleDto'
    SimpleRoleDto:
      type: object
      required:
        - id
        - name
        - isVisible
      properties:
        id:
          type: integer
          format: int64
          description: Role ID
        name:
          type: string
          description: Role name
        isVisible:
          type: boolean
          description: Is role visible on UI